+++ src/services/data.rs
@@ -1,5 +1,3 @@
 pub mod reputations;
 
-pub trait Data{
-    fn to_string(&self) -> Option<String>;
-}
+pub trait Data{}