+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- settings
+  - `bot_id` - loads `TELOXIDE_TOKEN` to env from settings
+  - `can_self_rep` - loads can user give rep to himself
+  - `can_rep_bot` - loads can user give rep to bot
+
 ## [0.1.0] - 2022-06-06
 
 ### Added